refactor(core): move pure markdown pipeline into @repo/core (phase 2)
Domain-code migration, phase 2: the pure remark/unified markdown pipeline moves
out of the desktop renderer into @repo/core so mobile can parse notes into mdast
+ gate the MDX vocabulary. The Plate/Slate serialization (md-rules.ts,
markdown-doc.ts) STAYS in the renderer and re-imports the pure pipeline from core.

Moved apps/desktop/src/renderer/editor/markdown/ → packages/core/src/markdown/:
  parse.ts (parseMdast), vocabulary.ts (the out-of-vocabulary gate),
  md-plugins.ts (MD_REMARK_PLUGINS + MD_STRINGIFY conventions), remark-mdx-agnostic.ts
Renderer markdown dir now holds only the two Plate-coupled files.

core gains runtime deps remark-math / mdast-util-mdx / micromark-extension-mdx /
mdast-util-gfm-task-list-item (all pure JS; the MDX ext is acorn-free) + promotes
mdast-util-to-markdown dev→dep, and exports ./markdown/{parse,vocabulary,md-plugins}.
Repointed markdown-doc, md-rules, markdown-kit + 2 tests to @repo/core/markdown/*.
Pruned 9 now-transitive markdown deps from apps/desktop.

Byte-stability held: the round-trip fixture matrix + kit-parity stay in the
renderer, exercise the moved pipeline through the full Plate path, and pass with
ZERO fixture bytes changed. core stays pure (typecheck under WebWorker/types:[]).
