+++ apps/desktop/package.json
@@ -62,11 +62,7 @@
     "katex": "^0.16.22",
     "lowlight": "^3.3.0",
     "lucide-react": "catalog:",
-    "mdast-util-gfm-task-list-item": "^2.0.0",
-    "mdast-util-mdx": "^3.0.0",
-    "mdast-util-to-markdown": "^2.1.2",
     "mermaid": "^11.16.0",
-    "micromark-extension-mdx": "^2.1.0",
     "nanoid": "^5.1.11",
     "platejs": "^53.2.4",
     "react": "catalog:",
@@ -74,13 +70,8 @@
     "react-dom": "catalog:",
     "react-lite-youtube-embed": "^3.6.0",
     "react-tweet": "^3.3.1",
-    "remark-frontmatter": "^5.0.0",
-    "remark-gfm": "^4.0.1",
-    "remark-math": "^6.0.0",
-    "remark-parse": "^11.0.0",
     "sherpa-onnx-node": "^1.10.41",
     "streamdown": "^2.5.0",
-    "unified": "^11.0.5",
     "use-stick-to-bottom": "^1.1.4",
     "zustand": "^5.0.13"
   },
