+++ packages/core/src/markdown/vocabulary.ts
